# Coalesce the Gateway.spec.servers[] with the same port and protocol